feat(frontend): usos cleanup
