+++ frontend/package.json
@@ -40,23 +40,19 @@
     "@types/canvas-confetti": "^1.9.0",
     "@types/uuid": "^10.0.0",
     "canvas-confetti": "^1.9.3",
-    "cheerio": "^1.0.0-rc.12",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "cmdk": "^1.0.0",
     "crypto-js": "^4.2.0",
     "date-fns": "^4.1.0",
-    "fetch-cookie": "^3.0.1",
     "framer-motion": "^11.3.28",
     "html-to-image": "^1.11.11",
     "input-otp": "^1.4.2",
     "jotai": "^2.9.3",
-    "lru-cache": "^11.0.1",
     "lucide-react": "^0.426.0",
     "next": "^15.1.7",
     "next-sitemap": "^4.2.3",
     "next-themes": "^0.4.3",
-    "node-fetch": "^3.3.2",
     "oauth-1.0a": "^2.2.6",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
