+++ frontend/src/app/plans/page.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import type { ExtendedCourse } from "@/atoms/plan-family";
 import { fetchToAdonis } from "@/lib/auth";
-import type { Registration } from "@/lib/types";
+import type { Registration } from "@/types";
 
 import { PlansPage } from "./page.client";
