--- frontend/src/lib/types.ts
@@ -1,43 +1,0 @@
-import type { ExtendedCourse } from "@/atoms/plan-family";
-import type { Day } from "@/services/usos/types";
-
-export interface ClassBlockProps {
-  startTime: string;
-  endTime: string;
-  groupId: string;
-  groupNumber: string;
-  groupOnlineId: number;
-  courseId: string;
-  courseName: string;
-  lecturer: string;
-  day: Day;
-  week: "" | "TN" | "TP";
-  courseType: "C" | "L" | "P" | "S" | "W";
-  registrationId: string;
-  spotsOccupied: number;
-  spotsTotal: number;
-  averageRating: number;
-  opinionsCount: number;
-}
-
-export interface Registration {
-  id: string;
-  name: string;
-}
-
-export interface Course {
-  name: string;
-  id: string;
-  type: string;
-  registrationId: string;
-}
-
-export interface SharedPlan {
-  id: string;
-  plan: {
-    name: string;
-    courses: ExtendedCourse[];
-    registrations: Registration[];
-    allGroups: ExtendedCourse["groups"];
-  };
-}
