--- frontend/src/services/usos/types.ts
@@ -1,25 +1,0 @@
-export enum Frequency {
-  EVERY = "każd",
-  ODD = "nieparzyst",
-  EVEN = "parzyst",
-  NEVER = "NEVER",
-}
-
-export enum Day {
-  MONDAY = "poniedziałek",
-  TUESDAY = "wtorek",
-  WEDNESDAY = "środa",
-  THURSDAY = "czwartek",
-  FRIDAY = "piątek",
-  SATURDAY = "sobota",
-  SUNDAY = "niedziela",
-  NEVER = "NEVER",
-}
-
-export enum LessonType {
-  LECTURE = "wyklad",
-  EXERCISES = "cwiczenia",
-  LABORATORY = "laboratorium",
-  PROJECT = "projekt",
-  SEMINAR = "seminarium",
-}
